pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "peakrdl-svd"
dynamic = ["version"]
requires-python = ">=3.9"
dependencies = [
    "systemrdl-compiler ~= 1.24",
    "cmsis-svd-codec"
]

authors = [
    {name="Alex Mykyta"},
    {name="Eyck Jentzsch"}
]
description = "Export SVD XML from the systemrdl-compiler register model"
readme = "README.md"
license = {text = "LGPLv3"}
keywords = [
    "SystemRDL", "PeakRDL", "CSR", "compiler", "tool", "registers", "generator",
    "SVD",
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU Lesser General Public License v3 (LGPLv3)",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Electronic Design Automation (EDA)",
]

[project.optional-dependencies]
cli = [
    "peakrdl-cli >= 1.2.3",
]

[project.urls]
Documentation = "https://peakrdl-svd.readthedocs.io"
Source = "https://github.com/SystemRDL/PeakRDL-svd"
Tracker = "https://github.com/SystemRDL/PeakRDL-svd/issues"
Changelog = "https://github.com/SystemRDL/PeakRDL-svd/releases"

[tool.setuptools.dynamic]
version = {attr = "peakrdl_svd.__about__.__version__"}

[project.entry-points."peakrdl.exporters"]
svd = "peakrdl_svd.__peakrdl__:Exporter"
